# Backend/docs/swagger.yaml
basePath: /api/v2/
definitions:
...
  entity.CommissionSettings:
    properties:
      cent:
        type: number
      percent:
        type: number
    type: object
...
  entity.MerchantCommission:
    properties:
      commission_active:
        type: boolean
      commission_cent:
        type: number
      commission_percent:
        type: number
      merchant_id:
        type: string
    type: object
...
host: 196.190.251.194:8082
info:
  contact:
    email: support@socialpay.com
    name: API Support
    url: http://www.swagger.io/support
  description: SocialPay API V2 documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SocialPay API V2
  version: "2.0"
paths:
...
  /admin/commissions/default:
    get:
      consumes:
      - application/json
      description: Retrieves the default commission settings for all merchants
      produces:
      - application/json
      responses:
        "200":
          description: 'settings: entity.CommissionSettings'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: error message'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get default commission settings
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Updates the default commission settings for all merchants
      parameters:
      - description: Commission Settings
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/entity.CommissionSettings'
      produces:
      - application/json
      responses:
        "200":
          description: 'settings: entity.CommissionSettings'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: invalid request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: error message'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update default commission settings
      tags:
      - admin
  /admin/commissions/merchant/{merchantID}:
    get:
      consumes:
      - application/json
      description: Retrieves the commission settings for a specific merchant
      parameters:
      - description: Merchant ID
        format: uuid
        in: path
        name: merchantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'commission: entity.MerchantCommission'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: invalid merchant ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: error message'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get merchant commission settings
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Updates the commission settings for a specific merchant
      parameters:
      - description: Merchant ID
        format: uuid
        in: path
        name: merchantID
        required: true
        type: string
      - description: Merchant Commission Settings
        in: body
        name: commission
        required: true
        schema:
          $ref: '#/definitions/entity.MerchantCommission'
      produces:
      - application/json
      responses:
        "200":
          description: 'commission: entity.MerchantCommission'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: invalid request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: error message'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update merchant commission settings
      tags:
      - admin
...
security:
- BearerAuth: []
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication for API key management endpoints
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Bearer token authentication for protected endpoints. Use format:
      "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
  MerchantID:
    description: Merchant ID authentication for merchant endpoints
    in: header
    name: X-MERCHANT-ID
    type: apiKey
swagger: "2.0"